Extract query repair command helpers partial
